refactor: address PR review — lift canvas setup, drop context layer
Replaces CanvasEditorContext (1-to-1 channel masquerading as context) with
prop-drilling from PostPromptViewInner. Restores VideoLayoutProvider's prop
interface to decouple video from canvas. Consolidates useTransitionType to a
single call site by threading transitionType as a prop. Reverts unrelated
style churn in Canvas activeElement memo. Keeps the useImperativeHandle
removal.
