+++ app/components/video/VideoLayoutContext.tsx
@@ -1,10 +1,12 @@
 "use client";
 
 import { createContext, use, useMemo, type ReactNode } from "react";
+import type { Editor } from "slate";
 import type { SceneElement } from "@/lib/canvas/types";
 import { useQueueSelector } from "@/lib/generation/GenerationQueueProvider";
+import type { AspectRatio } from "@/lib/video/aspectRatio";
+import type { TransitionType } from "@/lib/video/transitions";
 import type { VideoLayout } from "@/lib/video/types";
-import { useCanvasEditor } from "../canvas/CanvasEditorContext";
 import { useAssetPrefetch } from "./useAssetPrefetch";
 import { useVideoLayout } from "./useVideoLayout";
 
@@ -22,9 +24,25 @@ const VideoLayoutContext = createContext<VideoLayoutValue>({
 	scenes: [],
 });
 
-export function VideoLayoutProvider({ children }: { children: ReactNode }) {
-	const { editor, layoutKey } = useCanvasEditor();
-	const { layout, playerKey, scenes } = useVideoLayout(editor, layoutKey);
+export function VideoLayoutProvider({
+	editor,
+	layoutKey,
+	transitionType,
+	aspectRatio,
+	children,
+}: {
+	editor: Editor;
+	layoutKey: string;
+	transitionType: TransitionType;
+	aspectRatio: AspectRatio;
+	children: ReactNode;
+}) {
+	const { layout, playerKey, scenes } = useVideoLayout(
+		editor,
+		layoutKey,
+		transitionType,
+		aspectRatio,
+	);
 	const prefetched = useAssetPrefetch(layout);
 	const busy = useQueueSelector((q) => q.isBusy());
 	const ready = prefetched && !busy;
